make points given for won map and match configurable on post
Posts have to define mapPoints and matchPoints to configure how many points are
given in the ranking. 0 and 3 would give 3 points for winning the match, but no
points in the tournament ranking for just winning a map. 1 and 1 would give a
point for every map won and one point when winning the match. This is very
similar to the case before when playing best-of-three, only 2-1 would be
different as the loosing party still gets one tournament point.