let user customize colors
In the upper left is a details-tag containing color inputs which allow the user
to change all colors. The chosen colors are stored in the local storage and
restored from there on every page load. There's also a reset link, removing all
colors from the local storage and reloading the site.